implement rate action
