+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/about/AboutFragment.kt
@@ -1,6 +1,5 @@
 package au.com.shiftyjelly.pocketcasts.settings.about
 
-import android.content.ActivityNotFoundException
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
@@ -67,6 +66,7 @@ import au.com.shiftyjelly.pocketcasts.ui.extensions.getComposeThemeColor
 import au.com.shiftyjelly.pocketcasts.ui.extensions.getThemeDrawable
 import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.utils.extensions.pxToDp
+import au.com.shiftyjelly.pocketcasts.utils.rateUs
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import dagger.hilt.android.AndroidEntryPoint
@@ -382,19 +382,6 @@ private fun shareWithFriends(context: Context) {
         context.startActivity(Intent.createChooser(intent, context.getString(LR.string.share)))
     } catch (e: IllegalStateException) {
         // Not attached to activity anymore
-    }
-}
-
-private fun rateUs(context: Context) {
-    val packageName = context.packageName.removeSuffix(".debug")
-    val uri = Uri.parse("market://details?id=$packageName")
-    val goToMarket = Intent(Intent.ACTION_VIEW, uri)
-    // To count with Play market backstack, After pressing back button, to taken back to our application, we need to add following flags to intent.
-    goToMarket.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY or Intent.FLAG_ACTIVITY_NEW_DOCUMENT or Intent.FLAG_ACTIVITY_MULTIPLE_TASK)
-    try {
-        context.startActivity(goToMarket)
-    } catch (e: ActivityNotFoundException) {
-        context.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("http://play.google.com/store/apps/details?id=$packageName")))
     }
 }
